--- a/scripts/build/gulp/tasks/build.js
+++ b/scripts/build/gulp/tasks/build.js
@@ -6,19 +6,59 @@ const gulp = require('gulp');
 const path = require('path');
 const runSequence = require('run-sequence');
 const template = require('gulp-template');
+const webpack = require('webpack');
 const Constants = require('../constants');
 
-gulp.task('build', function(callback) {
-	const tasks = ['build-demo', 'post-cleanup'];
+runSequence.options.ignoreUndefinedTasks = true;
 
-	if (argv.release) {
-		tasks.push('minimize-css');
-	}
+function ifRelease(task) {
+	return !!argv.release && task;
+}
 
+function createBundle(configName, callback) {
+	const config = require(path.join(process.env.PWD, configName));
+	webpack(config, (error, {stats}) => {
+		if (error) {
+			callback(error);
+			return;
+		}
+		const errors = stats.reduce((acc, {compilation: {errors}}) => {
+			if (errors) {
+				acc.push(...errors);
+			}
+			return acc;
+		}, []);
+		const [firstError, ...remainingErrors] = errors;
+		if (firstError) {
+			remainingErrors.forEach(console.log.bind(console));
+			callback(firstError);
+		}
+		console.log(stats.toString({colors: true}));
+		callback();
+	});
+}
+
+gulp.task('webpack.prod', callback => {
+	createBundle('webpack.prod', callback);
+});
+
+gulp.task('webpack.dev', callback => {
+	createBundle('webpack.dev', callback);
+});
+
+gulp.task('build:assets', function(callback) {
 	runSequence(
 		'clean-dist',
 		['build-css', 'copy-ckeditor', 'copy-languages', 'copy-svgs'],
-		tasks,
+		['build-demo', 'post-cleanup', ifRelease('minimize-css')],
+		callback
+	);
+});
+
+gulp.task('build', function(callback) {
+	runSequence(
+		'build:assets',
+		[ifRelease('webpack.prod'), 'webpack.dev'],
 		callback
 	);
 });