--- a/webpack.common.js
+++ b/webpack.common.js
@@ -1,29 +1,51 @@
 const path = require('path');
 
-const modules = {
-	rules: [
-		{
-			test: /\.(js|jsx)$/,
-			exclude: /(node_modules|lib)/,
-			loader: 'babel-loader'
-		}
-	]
+/**
+ * Turns a root-relative path into an absolute one.
+ *
+ * This ensures that webpack works identically when invoked from the
+ * top-level (eg. with `npm run webpack` or from a subdirectory (eg.
+ * from a file under `scripts/gulp`).
+ */
+function toAbsolute(rootRelativePath) {
+	return path.join(__dirname, rootRelativePath);
+}
+
+const base = {
+	/**
+	 * https://webpack.js.org/configuration/entry-context/
+	 */
+	context: toAbsolute('.'),
+
+	/**
+	 * https://webpack.js.org/configuration/module/
+	 */
+	module: {
+		rules: [
+			{
+				test: /\.(js|jsx)$/,
+				exclude: /(node_modules|lib)/,
+				loader: 'babel-loader'
+			}
+		]
+	}
 };
 
 module.exports = {
 	config: {
+		...base,
 		entry: './scripts/build/index.js',
 		output: {
-			path: path.resolve('dist/alloy-editor')
-		},
-		module: modules
+			path: toAbsolute('./dist/alloy-editor')
+		}
 	},
 	core: {
+		...base,
 		entry: './src/adapter/main.js',
 		output: {
 			library: 'AlloyEditor',
 			libraryTarget: 'window',
-			path: path.resolve('dist/alloy-editor')
+			path: toAbsolute('./dist/alloy-editor')
 		},
 		externals: {
 			react: {
@@ -40,24 +62,24 @@ module.exports = {
 				amd: 'react-dom',
 				umd: 'react-dom'
 			}
-		},
-		module: modules
+		}
 	},
 	noCkeditor: {
+		...base,
 		entry: './src/adapter/main.js',
 		output: {
 			library: 'AlloyEditor',
 			libraryTarget: 'window',
-			path: path.resolve('dist/alloy-editor')
-		},
-		module: modules
+			path: toAbsolute('./dist/alloy-editor')
+		}
 	},
 	noReact: {
+		...base,
 		entry: './scripts/build/index.js',
 		output: {
 			library: 'AlloyEditor',
 			libraryTarget: 'umd',
-			path: path.resolve('dist/alloy-editor')
+			path: toAbsolute('./dist/alloy-editor')
 		},
 		externals: {
 			react: {
@@ -74,7 +96,6 @@ module.exports = {
 				amd: 'react-dom',
 				umd: 'react-dom'
 			}
-		},
-		module: modules
+		}
 	}
 };