--- a/webpack.dev.js
+++ b/webpack.dev.js
@@ -1,5 +1,4 @@
 const merge = require('webpack-merge');
-const webpack = require('webpack');
 
 const common = require('./webpack.common.js');