@wincent 2019-01-22
Rebased on top of #970, which means we will test the build process and the build result in Travis.