@wincent 2019-01-22
And rebased again on top op #971, which will increase our chances of avoiding spurious failures on Travis.